Relocated interfaces and their subclasses
