+++ objectRecognitionSource/src/main/java/io/vantiq/extsrc/objectRecognition/exception/FatalImageException.java
@@ -1,7 +1,7 @@
 package io.vantiq.extsrc.objectRecognition.exception;
 
-import io.vantiq.extsrc.objectRecognition.ImageRetrieverInterface;
-import io.vantiq.extsrc.objectRecognition.NeuralNetInterface;
+import io.vantiq.extsrc.objectRecognition.imageRetriever.ImageRetrieverInterface;
+import io.vantiq.extsrc.objectRecognition.neuralNet.NeuralNetInterface;
 
 /**
  * Indicates that an error occurred for either a {@link ImageRetrieverInterface} or a {@link NeuralNetInterface}
